update tests for new types
