+++ src/rmpr/types.ts
@@ -1,8 +1,7 @@
 import type { CommentNode, ReviewThreadNode } from '../common/github/pull_requests.ts';
 
 export interface DetailedReviewComment {
-  thread: ReviewThreadNode;
-  comment: CommentNode;
+  thread: Pick<ReviewThreadNode, 'id' | 'path' | 'originalLine' | 'originalStartLine'>;
+  comment: Pick<CommentNode, 'id' | 'body' | 'diffHunk' | 'author'>;
   diffForContext: string;
-  diffForTerminal: string;
 }
